restructured

+++ src/code/systems/mongoTs/generators/init.ts
@@ -8,14 +8,20 @@ const createUtils = () => {
   if (!fs.existsSync(srcdir)) {
     fs.mkdirSync(srcdir, { recursive: true });
   }
-  fs.writeFileSync(path.join(srcdir, 'Utils.ts'), utils.ts);
+  const utilsPath = path.join(srcdir, 'Utils.ts');
+  if (!fs.existsSync(utilsPath)) {
+    fs.writeFileSync(utilsPath, utils.ts);
+  }
 };
 const createDbAndDbDir = () => {
   const dbDir = path.join(Config.projectPath, Config.get('srcdir'), 'db');
   if (!fs.existsSync(dbDir)) {
     fs.mkdirSync(dbDir, { recursive: true });
   }
-  fs.writeFileSync(path.join(dbDir, 'mongo.ts'), db({ name: Config.get('name') }).ts);
+  const dbPath = path.join(dbDir, 'mongo.ts');
+  if (!fs.existsSync(dbPath)) {
+    fs.writeFileSync(path.join(dbDir, 'mongo.ts'), db({ name: Config.get('name') }).ts);
+  }
 };
 
 export const init = () => {
